Create parser add abs function
