+++ README.md
@@ -4,3 +4,32 @@
 
 
 The project is built based on UIUC CS-421 MP2 (https://github-dev.cs.illinois.edu/cs421-su21/release.git)
+
+
+## Running Code
+Install Haskell 
+
+To run the interpolator, start GHCi with `stack ghci`. From here, run REPL by calling `main`.
+
+Type Python commands under Python REPL:
+
+`Python>`
+
+Use `quit` or cmd + c to quit the interpolator.
+
+## Interpolator Functions
+
+### Constants
+Recognizable constants include
+* integer
+* string
+
+### Variables
+Assign a constant to a variable
+
+```
+Python> a = 3
+
+Python> print(a)
+3
+```
