+++ interpreter.cabal
@@ -31,6 +31,7 @@ library
 executable interpreter
   main-is: Main.hs
   other-modules:
+      Parse
       Paths_interpreter
   hs-source-dirs:
       app
